+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ### Fixed
 
 - Fix JDK folder detection on Linux and improve JDK version handling.
+- Fix build on NixOS.
 
 ### Security